+++ CustomerMoghimiHome/Server/EntityFramework/IRepository.cs
@@ -1,8 +1,23 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
-
 namespace CustomerMoghimiHome.Shared.EntityFramework.Common;
 
+
+public interface IRepository<TEntity> where TEntity : class, IBaseEntity
+{
+    Task<bool> ExistsAsync(int id);
+    Task<bool> ExistsAsync(Expression<Func<TEntity, bool>> predicate);
+
+    void Add(TEntity entity);
+    void AddRange(IEnumerable<TEntity> entities);
+    void Remove(TEntity entity);
+    void RemoveRange(IEnumerable<TEntity> entities);
+    void Update(TEntity entity);
+    void UpdateRange(IEnumerable<TEntity> entities);
+
+    Task AddAsync(TEntity entity);
+    Task AddRangeAsync(IEnumerable<TEntity> entities);
+}
 
 public class Repository<TEntity> : IRepository<TEntity> where TEntity : class, IBaseEntity
 {
