+++ CustomerMoghimiHome/Server/EntityFramework/IUnitOfWork.cs
@@ -1,4 +1,11 @@
 ﻿namespace CustomerMoghimiHome.Shared.EntityFramework.Common;
+public interface IUnitOfWork : IDisposable
+{
+    //IPostRepository Posts { get; }
+    //IPostCategoryRepository PostCategories { get; }
+    //IPostFeedBackRepository PostFeedBacks { get; }
+    Task<bool> CommitAsync();
+}
 
 public class UnitOfWork //: IUnitOfWork
 {
