--- CustomerMoghimiHome/Shared/EntityFramework/Common/IUnitOfWork.cs
@@ -1,8 +1,0 @@
-﻿namespace CustomerMoghimiHome.Shared.EntityFramework.Common;
-public interface IUnitOfWork : IDisposable
-{
-    //IPostRepository Posts { get; }
-    //IPostCategoryRepository PostCategories { get; }
-    //IPostFeedBackRepository PostFeedBacks { get; }
-    Task<bool> CommitAsync();
-}
